Required changes made
