+++ sympy/matrices/tests/test_matrices.py
@@ -2615,45 +2615,6 @@ def test_simplify_immutable():
     assert simplify(ImmutableMatrix([[sin(x)**2 + cos(x)**2]])) == \
                     ImmutableMatrix([[1]])
 
-def test_rank():
-    from sympy.abc import x
-    m = Matrix([[1, 2], [x, 1 - 1/x]])
-    assert m.rank() == 2
-    n = Matrix(3, 3, range(1, 10))
-    assert n.rank() == 2
-    p = zeros(3)
-    assert p.rank() == 0
-
-def test_issue_11434():
-    ax, ay, bx, by, cx, cy, dx, dy, ex, ey, t0, t1 = \
-        symbols('a_x a_y b_x b_y c_x c_y d_x d_y e_x e_y t_0 t_1')
-    M = Matrix([[ax, ay, ax*t0, ay*t0, 0],
-                [bx, by, bx*t0, by*t0, 0],
-                [cx, cy, cx*t0, cy*t0, 1],
-                [dx, dy, dx*t0, dy*t0, 1],
-                [ex, ey, 2*ex*t1 - ex*t0, 2*ey*t1 - ey*t0, 0]])
-    assert M.rank() == 4
-
-def test_rank_regression_from_so():
-    # see:
-    # https://stackoverflow.com/questions/19072700/why-does-sympy-give-me-the-wrong-answer-when-i-row-reduce-a-symbolic-matrix
-
-    nu, lamb = symbols('nu, lambda')
-    A = Matrix([[-3*nu,         1,                  0,  0],
-                [ 3*nu, -2*nu - 1,                  2,  0],
-                [    0,      2*nu, (-1*nu) - lamb - 2,  3],
-                [    0,         0,          nu + lamb, -3]])
-    expected_reduced = Matrix([[1, 0, 0, 1/(nu**2*(-lamb - nu))],
-                               [0, 1, 0,    3/(nu*(-lamb - nu))],
-                               [0, 0, 1,         3/(-lamb - nu)],
-                               [0, 0, 0,                      0]])
-    expected_pivots = (0, 1, 2)
-
-    reduced, pivots = A.rref()
-
-    assert simplify(expected_reduced - reduced) == zeros(*A.shape)
-    assert pivots == expected_pivots
-
 def test_replace():
     from sympy import symbols, Function, Matrix
     F, G = symbols('F, G', cls=Function)
@@ -2997,13 +2958,6 @@ def test_case_6913():
     a = m[0, 0]>0
     assert str(a) == 'm[0, 0] > 0'
 
-def test_issue_15872():
-    A = Matrix([[1, 1, 1, 0], [-2, -1, 0, -1], [0, 0, -1, -1], [0, 0, 2, 1]])
-    B = A - Matrix.eye(4) * I
-    assert B.rank() == 3
-    assert (B**2).rank() == 2
-    assert (B**3).rank() == 2
-
 def test_issue_11948():
     A = MatrixSymbol('A', 3, 3)
     a = Wild('a')
@@ -3017,27 +2971,6 @@ def test_gramschmidt_conjugate_dot():
     mat = Matrix([[1, I], [1, -I]])
     Q, R = mat.QRdecomposition()
     assert Q * Q.H == Matrix.eye(2)
-
-def test_issue_17827():
-    C = Matrix([
-        [3, 4, -1, 1],
-        [9, 12, -3, 3],
-        [0, 2, 1, 3],
-        [2, 3, 0, -2],
-        [0, 3, 3, -5],
-        [8, 15, 0, 6]
-    ])
-    # Tests for row/col within valid range
-    D = C.elementary_row_op('n<->m', row1=2, row2=5)
-    E = C.elementary_row_op('n->n+km', row1=5, row2=3, k=-4)
-    F = C.elementary_row_op('n->kn', row=5, k=2)
-    assert(D[5, :] == Matrix([[0, 2, 1, 3]]))
-    assert(E[5, :] == Matrix([[0, 3, 0, 14]]))
-    assert(F[5, :] == Matrix([[16, 30, 0, 12]]))
-    # Tests for row/col out of range
-    raises(ValueError, lambda: C.elementary_row_op('n<->m', row1=2, row2=6))
-    raises(ValueError, lambda: C.elementary_row_op('n->kn', row=7, k=2))
-    raises(ValueError, lambda: C.elementary_row_op('n->n+km', row1=-1, row2=5, k=2))
 
 def test_issue_8207():
     a = Matrix(MatrixSymbol('a', 3, 1))
